requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database and ORM
supabase==2.0.2
postgrest==0.13.2

# HTTP requests
requests==2.31.0
httpx==0.24.1

# Environment and configuration
python-dotenv==1.0.0

# Security and authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Optional: Redis for caching (uncomment if using Redis)
# redis==5.0.1

# Optional: Celery for background tasks (uncomment if using Celery)
# celery==5.3.4
